🔧 refactor(apiSchemas.ts): replace generic Record types with specific DTOs for better type safety ✨ feat(stats.tsx, layout.tsx): add new VehiculeStats component and layout for displaying vehicle statistics 🔧 refactor(page.tsx): integrate VehiculeStats into the page and improve layout
The VehicleStatsDTO and VehicleResponseDTO in apiSchemas.ts were previously defined as generic Record types, which provided no type safety. They have been replaced with specific DTOs to improve type safety and readability.

A new VehiculeStats component has been added to display vehicle statistics. This component fetches data from the API and displays it in a series of SimpleStatisticCards. This provides a more detailed and visually appealing overview of vehicle statistics.

The layout of the page.tsx file has been refactored to integrate the new VehiculeStats component and improve the overall layout. The title has been moved to a new layout.tsx file to separate concerns and improve readability.