🐛 fix(VehicleResponseDTO): change estResrvee type from boolean to string 🔧 refactor(page.tsx): replace VehiculeComponent with ClientPaginatedTable for better data presentation
The type of estResrvee in VehicleResponseDTO was changed from boolean to string to match the data type received from the backend. This ensures that the data is correctly interpreted and displayed.

The VehiculeComponent in page.tsx was replaced with ClientPaginatedTable to improve the presentation of data. This change provides a more structured and readable format for displaying vehicle data, enhancing the user experience. The RenderCell function was added to customize the rendering of each cell in the table based on the columnKey.